% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{A function to tell you when it's someone's birthday!}
\usage{
birthday(Name, Month, Day)
}
\arguments{
\item{Name}{name of the individual}

\item{Month}{their birthday month ("July")}

\item{Day}{the day of their birthday}
}
\description{
A function to tell you when it's someone's birthday!
}
\examples{
birthday("Marina", "July", 16)
}
